refactor: extract FundamentalsPage components + formatters (F5)
Extract the six presentational components (FinancialTable, MetricCard,
TabButton, SentimentBadge, ResearchProfileList, AnnouncementsPanel) into
components/fundamentals/components.js and the shared formatters/tone map/
section-heading style (fmtBig, currencyPrefix, metricToneClass,
sectionTitleClass) into components/fundamentals/format.js. The page keeps
its fetch/state logic and its main-only data (INCOME_ROWS/BALANCE_ROWS/
CASHFLOW_ROWS, tab configs) and drops to 863 lines. Pure code movement,
guarded by FundamentalsPage's existing tests; lint + build clean.
